+++ packages/atomic/src/components/result-lists/atomic-folded-result-list/atomic-folded-result-list.tsx
@@ -8,44 +8,43 @@ import {
   FoldedResultListState,
   buildResultsPerPage,
   ResultListProps,
+  FoldedCollection,
 } from '@coveo/headless';
 import {
   Bindings,
   BindStateToController,
   InitializableComponent,
   InitializeBindings,
-} from '../../utils/initialization-utils';
+} from '../../../utils/initialization-utils';
 import {
   ResultDisplayDensity,
   ResultDisplayImageSize,
-} from '../atomic-result/atomic-result-display-options';
-import {
-  ResultListCommon,
-  ResultRenderingFunction,
-} from '../result-lists/result-list-common';
+} from '../../atomic-result/atomic-result-display-options';
+import {ResultListCommon, ResultRenderingFunction} from '../result-list-common';
+import {FoldedResultListStateContextEvent} from '../result-list-decorators';
 
 /**
- * The `atomic-folded-result-list` component is responsible for displaying query results by applying one or more result templates.
+ * The `atomic-folded-result-list` component is responsible for displaying folded query results, by applying one or more result templates for up to three layers (i.e., to the result, child and grandchild).
  *
  * @internal
  * @part result-list - The element containing every result of a result list
  */
 @Component({
   tag: 'atomic-folded-result-list',
-  styleUrl: 'atomic-folded-result-list.pcss',
+  styleUrl: '../result-list-common.pcss',
   shadow: true,
 })
 export class AtomicFoldedResultList implements InitializableComponent {
   @InitializeBindings() public bindings!: Bindings;
-  public resultList!: FoldedResultList;
+  public foldedResultList!: FoldedResultList;
   public resultsPerPage!: ResultsPerPage;
   public listWrapperRef?: HTMLDivElement;
 
   @Element() public host!: HTMLDivElement;
 
-  @BindStateToController('resultList')
+  @BindStateToController('foldedResultList')
   @State()
-  public resultListState!: FoldedResultListState;
+  public foldedResultListState!: FoldedResultListState;
 
   @BindStateToController('resultsPerPage')
   @State()
@@ -58,7 +57,6 @@ export class AtomicFoldedResultList implements InitializableComponent {
 
   public resultListCommon!: ResultListCommon;
   /**
-   * TODO: KIT-452 Infinite scroll feature
    * Whether to automatically retrieve an additional page of results and append it to the
    * current results when the user scrolls down to the bottom of element
    */
@@ -77,15 +75,21 @@ export class AtomicFoldedResultList implements InitializableComponent {
    */
   @Prop({reflect: true}) imageSize: ResultDisplayImageSize = 'icon';
   /**
-   * TODO:
+   * The name of the field on which to do the folding. The folded result list component will use the values of this field to resolve the collections of result items.
+   *
+   * @defaultValue `foldingcollection`
    */
   @Prop({reflect: true}) public collectionField?: string;
   /**
-   * TODO:
+   * The name of the field that determines whether a certain result is a top result containing other child results within a collection.
+   *
+   * @defaultValue `foldingparent`
    */
   @Prop({reflect: true}) public parentField?: string;
   /**
-   * TODO:
+   * The name of the field that uniquely identifies a result within a collection.
+   *
+   * @defaultValue `foldingchild`
    */
   @Prop({reflect: true}) public childField?: string;
 
@@ -109,8 +113,22 @@ export class AtomicFoldedResultList implements InitializableComponent {
       this.enableInfiniteScroll &&
       this.resultListCommon.scrollHasReachedEndOfList
     ) {
-      this.resultList.fetchMoreResults();
+      this.foldedResultList.fetchMoreResults();
     }
+  }
+
+  @Listen('atomic/resolveFoldedResultList')
+  resolveFoldedResultList(event: FoldedResultListStateContextEvent) {
+    event.preventDefault();
+    event.stopPropagation();
+    event.detail(this.foldedResultList);
+  }
+
+  @Listen('atomic/loadCollection')
+  loadCollection(event: CustomEvent<FoldedCollection>) {
+    event.preventDefault();
+    event.stopPropagation();
+    this.foldedResultList.loadCollection(event.detail);
   }
 
   public async initialize() {
@@ -127,7 +145,7 @@ export class AtomicFoldedResultList implements InitializableComponent {
       },
     });
 
-    this.resultList = this.initFolding(
+    this.foldedResultList = this.initFolding(
       this.resultListCommon.resultListControllerProps
     );
     this.resultsPerPage = buildResultsPerPage(this.bindings.engine);
@@ -156,7 +174,7 @@ export class AtomicFoldedResultList implements InitializableComponent {
 
   public componentDidRender() {
     this.resultListCommon.componentDidRender(
-      this.resultListState.firstSearchExecuted,
+      this.foldedResultListState.firstSearchExecuted,
       this.listWrapperRef
     );
   }
@@ -167,7 +185,7 @@ export class AtomicFoldedResultList implements InitializableComponent {
       density: this.density,
       imageSize: this.imageSize,
       templateHasError: this.templateHasError,
-      resultListState: this.resultListState,
+      resultListState: this.foldedResultListState,
       resultsPerPageState: this.resultsPerPageState,
       setListWrapperRef: (el) => {
         this.listWrapperRef = el as HTMLDivElement;