+++ packages/atomic/src/components/result-templates/atomic-result-children-template/atomic-result-children-template.tsx
@@ -1,21 +1,22 @@
-import {Component, Element, Prop, Method, State} from '@stencil/core';
 import {ResultTemplate, ResultTemplateCondition} from '@coveo/headless';
-import {MapProp} from '../../utils/props-utils';
-import {ResultTemplateCommon} from '../result-templates/result-template-common';
+import {Component, Element, Prop, State, Method} from '@stencil/core';
+import {MapProp} from '../../../utils/props-utils';
+import {ResultTemplateCommon} from '../result-template-common';
 
 /**
- * The `atomic-result-template` component determines the format of the query results, depending on the conditions that are defined for each template. A `template` element must be the child of an `atomic-result-template`, and an `atomic-result-list` must be the parent of each `atomic-result-template`.
+ * The `atomic-result-children-template` component determines the format of the child results, depending on the conditions that are defined for each template. A `template` element must be the child of an `atomic-result-children-template`, and an `atomic-result-children` must be the parent of each `atomic-result-children-template`.
  *
  * Note: Any `<script>` tags defined inside of a `<template>` element will not be executed when results are being rendered.
+ * @internal
  */
 @Component({
-  tag: 'atomic-result-template',
+  tag: 'atomic-result-children-template',
   shadow: true,
 })
-export class AtomicResultTemplate {
-  @State() public error!: Error;
-
+export class AtomicResultChildrenTemplate {
   @Element() public host!: HTMLDivElement;
+
+  @State() public error!: Error;
 
   /**
    * A function that must return true on results for the result template to apply.
@@ -26,14 +27,14 @@ export class AtomicResultTemplate {
   @Prop() public conditions: ResultTemplateCondition[] = [];
 
   /**
-   * The field and values that define which result items the condition must be applied to.
+   * Fields and field values that results must match for the result template to apply.
    *
    * For example, a template with the following attribute only applies to result items whose `filetype` is `lithiummessage` or `YouTubePlaylist`: `must-match-filetype="lithiummessage,YouTubePlaylist"`
    */
   @MapProp({splitValues: true}) public mustMatch: Record<string, string[]> = {};
 
   /**
-   * The field and values that define which result items the condition must not be applied to.
+   * Fields and field values that results must not match for the result template to apply.
    *
    * For example, a template with the following attribute only applies to result items whose `filetype` is not `lithiummessage`: `must-not-match-filetype="lithiummessage"`
    */
@@ -42,21 +43,13 @@ export class AtomicResultTemplate {
 
   public resultTemplateCommon: ResultTemplateCommon;
 
-  public componentWillLoad() {
-    this.resultTemplateCommon.addMatchConditions(
-      this.mustMatch,
-      this.mustNotMatch
-    );
-  }
-
   constructor() {
     this.resultTemplateCommon = new ResultTemplateCommon({
       host: this.host,
       setError: (err) => {
         this.error = err;
       },
-      validParents: ['atomic-result-list', 'atomic-folded-result-list'],
-      allowEmpty: true,
+      validParents: ['atomic-result-children'],
     });
   }
 
@@ -66,6 +59,13 @@ export class AtomicResultTemplate {
   @Method()
   public async getTemplate(): Promise<ResultTemplate<DocumentFragment> | null> {
     return this.resultTemplateCommon.getTemplate(this.conditions, this.error);
+  }
+
+  public componentWillLoad() {
+    this.resultTemplateCommon.addMatchConditions(
+      this.mustMatch,
+      this.mustNotMatch
+    );
   }
 
   public render() {